Update.
	* sysdeps/unix/sysv/linux/sigwait.c: Include string.h.
	* sysdeps/unix/sysv/linux/sigwaitinfo.c: Likewise.
	* sysdeps/unix/sysv/linux/sigtimedwait.c: Likewise.
	* sysdeps/unix/sysv/linux/sleep.c (__sleep): Cast value to unsigned
	int before assigning to max to avoid warnings.

2003-07-11  Jakub Jelinek  <jakub@redhat.com>

+++ sysdeps/unix/sysv/linux/sleep.c
@@ -40,7 +40,8 @@ cl (void *arg)
 unsigned int
 __sleep (unsigned int seconds)
 {
-  const unsigned int max = ((unsigned long int) (~((time_t) 0))) >> 1;
+  const unsigned int max
+    = (unsigned int) (((unsigned long int) (~((time_t) 0))) >> 1);
   struct timespec ts;
   sigset_t set, oset;
   unsigned int result;
